requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
openai==1.58.1
python-telegram-bot==21.9
pydantic-settings==2.7.1
